requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pillow==11.1.0
python-multipart==0.0.20
openai==1.82.0
httpx==0.28.1
